Refactor sparse_dense_shout module: Enable sparse_dense_shout subprotocol and enhance sumcheck computations
- Un-commented the sparse_dense_shout module in mod.rs to enable its usage.
- Updated imports in sparse_dense_shout.rs to include necessary types from execution_trace.
- Removed redundant functions and code related to sumcheck computations, consolidating logic for clarity.
- Enhanced the prove_sparse_dense_shout function to utilize JoltONNXCycle and adjusted tensor size calculations.
- Improved parallel processing of lookup indices and operand evaluations for performance.
- Added back the current_suffix_len and compute_sumcheck_prover_message functions for clarity and reusability.
- Updated tests to accommodate changes in the execution trace structure and ensure correctness.
