+++ zkml-jolt-core/src/subprotocols/mod.rs
@@ -1,2 +1,2 @@
-// pub mod sparse_dense_shout;
+pub mod sparse_dense_shout;
 pub mod twist;
